Remove mux dependency from the registry client.
By moving functions that are only used by the handlers to
the registry/handlers package.
