+++ context/http_test.go
@@ -5,7 +5,6 @@ import (
 	"net/http/httptest"
 	"net/http/httputil"
 	"net/url"
-	"reflect"
 	"testing"
 	"time"
 )
@@ -176,47 +175,6 @@ func TestWithResponseWriter(t *testing.T) {
 
 	if ctx.Value("http.response.status") != http.StatusBadRequest {
 		t.Fatalf("unexpected response status in context: %v != %v", ctx.Value("http.response.status"), http.StatusBadRequest)
-	}
-}
-
-func TestWithVars(t *testing.T) {
-	var req http.Request
-	vars := map[string]string{
-		"foo": "asdf",
-		"bar": "qwer",
-	}
-
-	getVarsFromRequest = func(r *http.Request) map[string]string {
-		if r != &req {
-			t.Fatalf("unexpected request: %v != %v", r, req)
-		}
-
-		return vars
-	}
-
-	ctx := WithVars(Background(), &req)
-	for _, testcase := range []struct {
-		key      string
-		expected interface{}
-	}{
-		{
-			key:      "vars",
-			expected: vars,
-		},
-		{
-			key:      "vars.foo",
-			expected: "asdf",
-		},
-		{
-			key:      "vars.bar",
-			expected: "qwer",
-		},
-	} {
-		v := ctx.Value(testcase.key)
-
-		if !reflect.DeepEqual(v, testcase.expected) {
-			t.Fatalf("%q: %v != %v", testcase.key, v, testcase.expected)
-		}
 	}
 }
